refactor: User 관련 Entity 디렉토리 변경(domain)
